+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Fixed
 
 - **Projectroles**
     - Browser-specific ``sodar-btn-submit-once`` spinner padding (#1291)
+    - Hidden JSON app settings reset on non-superuser project update (#1295)
 
 
 v0.13.2 (2023-09-21)